Added previous get changes to PR
